+++ src/plugins/essentials/lyapunov.jl
@@ -9,7 +9,7 @@ struct Lyapunov <: AbstractPlugin
     ni::Int
     ts::Float64
 end
-Lyapunov(;m=15, J=5, ni=300, ts=0.01) = Lyapunov(m, J, ni, ts)
+Lyapunov(;m::Int=15, J::Int=5, ni::Int=300, ts::Float64=0.01) = Lyapunov(m, J, ni, ts)
 
 show(io::IO, plg::Lyapunov) = print(io, 
     "Lyapunov(embeddingdim:$(plg.m), numlags:$(plg.J), numiteration:$(plg.ni), samplingtime:$(plg.ts)")
