+++ src/plugins/essentials/mean.jl
@@ -4,7 +4,7 @@ using Statistics
 struct Mean <: AbstractPlugin
     dims::Int
 end
-Mean(;dims=1) = Mean(dims)
+Mean(;dims::Int=1) = Mean(dims)
 
 show(io::IO, plg::Mean) = print(io, "Mean(dims:$(plg.dims))")
